Make LoggingManagent::CreateLoggingManagement() API clearer and more robust
Previously, the API (and associated set of constructors) took a series
of arrays to initialize the LoggingManagement object.  All arrays were
supposed to be provided in the order from Debug to CriticalProduction
importance.  The code required 5 different arrays.  The ordering
required was the opposite of the numerical ordering of the importance
enum.  This led to confusion.  The new API replaces the prior arrays
with a single array of objects, each object specifies all required
storage and associated objects per importance level.  The API
documentation unabiguously states that the objects are passed in the
numerical order of importance.  Finally, the API is friendly
towards flash based structures, which should allow constrained
platforms to create the initialization arguments in flash.